Started Tag Requirements for Polls
Polls now have a way to select which tags can reply, whether the student has those tags in their array of tags, or if the student has the exact same tags chosen for the poll. Next step to do is send it to the students who's tags match the required tags.