Started a Timer Feature
Created buttons to start and stop a timer with a specified time in minutes and an option to play sound when the timer ends. The data is stored in classData, and can be used in app.js to transmit data between sockets. Pushing this for someone else to take and finish