Digipog Framework
Digipogs can be used in an app to make purchases.
The owner of the app will make half the digipogs they charged.
If the full flag is true, the owner of the app will make all of the digipogs they charged.
Digipogs transactions are tracked in a transaction table.
Apps allowed to use digipogs are tracked in a apps table.
Users have a digipog column in the users table.